feat(campus): migrate events / clubs / dining to hybrid SSR + SWR
Rolls the news pilot pattern out to the three remaining explore
surfaces using the API routes + hooks + skeletons added in the
previous commit.

New client renderers (all `"use client"`):

  EventsListClient
    Takes initial DB events (SWR-seeded) + ICS events (static).
    Localises DB rows then `mergeEvents` with ICS in a memo so a
    SWR revalidation only swaps the affected rows. Skeleton on
    cold cache.

  ClubsListClient
    Identical pattern, no merge complexity — straight SWR over
    the clubs response.

  DiningListClient
    Same again, with the gradient-image fallback inside the card.

Server pages
  Drop the inline JSX + local `BANNER_BG / AVATAR_BG` maps
  (centralised inside the client renderers), keep the SSR data
  fetch + ICS merge upstream of the client mount. First paint
  is real content via `fallbackData`; navigations within the SPA
  session pop straight from SWR cache; backgrounds revalidate
  on focus / reconnect.

Typecheck + lint clean.
