+++ apps/campus/app/(public)/campus/[token]/clubs/page.tsx
@@ -1,13 +1,12 @@
 import type { Metadata } from "next";
-import Link from "next/link";
 import { notFound } from "next/navigation";
-import { ExternalLink, MapPin } from "lucide-react";
 import { getPublicCampusByToken } from "@/lib/public-campus";
-import { detectLocale, pickLocalized } from "@/app/lib/i18n-core";
+import { detectLocale } from "@/app/lib/i18n-core";
 import { listTopClubsForProject } from "@/lib/clubs-db";
 import { ConsumerNav } from "@/lib/consumer/ConsumerNav";
 import { SegmentedTabs } from "@/lib/consumer/SegmentedTabs";
 import { ConsumerFooter } from "@/lib/consumer/ConsumerFooter";
+import { ClubsListClient } from "@/lib/consumer/ClubsListClient";
 
 type Params = Promise<{ token: string }>;
 
@@ -20,13 +19,6 @@ interface CampusBranding {
 function isValidHex(value: string | undefined): value is string {
   return !!value && /^#([0-9a-f]{3}|[0-9a-f]{6})$/i.test(value);
 }
-
-const AVATAR_BG: Record<string, string> = {
-  purple: "var(--brand-primary-fill)",
-  coral: "var(--brand-accent-warm)",
-  teal: "var(--brand-accent-cool)",
-  pink: "var(--brand-accent-complement)",
-};
 
 export async function generateMetadata({
   params,
@@ -103,91 +95,13 @@ export default async function ClubsPage({
           activity.
         </p>
 
-        {clubs.length === 0 ? (
-          <div className="mt-10 rounded-2xl border border-[var(--brand-line)] bg-white p-8 text-center text-sm text-[var(--brand-text-muted)]">
-            No clubs published yet.
-          </div>
-        ) : (
-          <div className="mt-8 grid grid-cols-1 gap-5 md:grid-cols-2 lg:grid-cols-3">
-            {clubs.map((c) => {
-              const name = pickLocalized(c.name, c.nameEl, locale);
-              const description = pickLocalized(
-                c.description,
-                c.descriptionEl,
-                locale,
-              );
-              return (
-              <article
-                key={c.id}
-                className="flex flex-col gap-3 rounded-2xl border border-[var(--brand-line)] bg-white p-5"
-              >
-                <div className="flex items-center gap-3">
-                  {c.imageUrl ? (
-                    // eslint-disable-next-line @next/next/no-img-element
-                    <img
-                      src={c.imageUrl}
-                      alt=""
-                      className="h-12 w-12 rounded-xl object-cover"
-                    />
-                  ) : (
-                    <span
-                      aria-hidden
-                      className="flex h-12 w-12 items-center justify-center rounded-xl text-sm font-medium text-white"
-                      style={{ backgroundColor: AVATAR_BG[c.avatarColor] }}
-                    >
-                      {c.initials}
-                    </span>
-                  )}
-                  <div className="min-w-0 flex-1">
-                    <Link
-                      href={`/campus/${token}/clubs/${c.id}${lang}`}
-                      className="block truncate text-base font-medium text-[var(--brand-text)] transition-colors hover:text-[var(--brand-primary)]"
-                    >
-                      {name}
-                    </Link>
-                    <p className="mt-0.5 truncate text-xs text-[var(--brand-text-muted)]">
-                      {c.memberCount} members
-                      {c.meetsCadence ? ` · ${c.meetsCadence}` : ""}
-                    </p>
-                  </div>
-                </div>
-
-                <p className="line-clamp-3 text-sm leading-relaxed text-[var(--brand-text)]">
-                  {description}
-                </p>
-
-                {c.anchors[0] ? (
-                  <span className="inline-flex w-fit items-center gap-1.5 rounded-full bg-[var(--brand-primary-bg)] px-3 py-1 text-xs font-medium text-[var(--brand-primary)]">
-                    <MapPin size={12} strokeWidth={1.75} />
-                    {c.anchors[0].refName}
-                  </span>
-                ) : null}
-
-                <div className="mt-auto flex items-center gap-2 pt-2">
-                  <Link
-                    href={`/campus/${token}/clubs/${c.id}${lang}`}
-                    className="rounded-full border border-[var(--brand-line)] bg-white px-3 py-1.5 text-xs font-medium text-[var(--brand-text)] transition-colors hover:border-[var(--brand-primary)]"
-                  >
-                    Details
-                  </Link>
-                  {c.externalLink ? (
-                    <a
-                      href={c.externalLink}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="inline-flex items-center gap-1 rounded-full px-3 py-1.5 text-xs font-medium text-white transition-opacity hover:opacity-90"
-                      style={{ backgroundColor: "var(--brand-primary)" }}
-                    >
-                      <ExternalLink size={12} strokeWidth={1.75} />
-                      View
-                    </a>
-                  ) : null}
-                </div>
-              </article>
-              );
-            })}
-          </div>
-        )}
+        <ClubsListClient
+          token={token}
+          locale={locale}
+          lang={lang}
+          initialClubs={clubs}
+          emptyCopy="No clubs published yet."
+        />
       </section>
 
       <ConsumerFooter campusName={campusName} />
